pyproject.toml
[project]
name = "apache-airflow-study"
version = "0.1.0"
description = ""
authors = [
    {name = "Phuc Huynh Van",email = "phuc.huynhvanspkt@gmail.com"}
]
readme = "README.md"
requires-python =  ">=3.10,<3.13"
dependencies = [
    "apache-airflow (>=2.10.5,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pylint (>=3.3.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
